+++ noderedpy/__init__.py
@@ -1,13 +1,19 @@
 # -*- coding: utf-8 -*-
-from ._nodered import RED, Node, NodeProperty
+from ._nodered import RED, Node
+from ._property import (
+    InputProperty, ListProperty, DictProperty,
+    SpinnerProperty, ComboBoxProperty
+)
 from ._server import Server, StandaloneServer
 
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 __author__ = "oyajiDev"
 __email__ = "this.dev.somehit@gmail.com"
 
 __all__ = [
-    "RED", "Node", "NodeProperty",
+    "RED", "Node",
+    "InputProperty", "ListProperty", "DictProperty",
+    "SpinnerProperty", "ComboBoxProperty",
     "Server", "StandaloneServer"
 ]