+++ tests/server_test.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
 import os
-from noderedpy import NodeProperty, RED, Server
+from noderedpy import (
+    RED, Server,
+    InputProperty, ListProperty, DictProperty,
+    SpinnerProperty, ComboBoxProperty
+)
 from noderedpy.decorator import register
 
 __dirname = os.path.dirname(os.path.realpath(__file__))
@@ -9,9 +13,9 @@ if __name__ == "__main__":
         count = 0
 
         @register("test", properties = [
-            NodeProperty("test_prop", "str", "1234"),
-            NodeProperty("list_prop", "list", [ 1, 2, 3, 4 ]),
-            NodeProperty("dict_prop", "dict", { "a": 1 })
+            InputProperty("test_prop", "1234"),
+            ListProperty("list_prop", [ 1, 2, 3, 4 ]),
+            DictProperty("dict_prop", { "a": 1 })
         ])
         def test(props:dict, msg:dict) -> dict:
             TotalApp.count += 1
@@ -21,6 +25,17 @@ if __name__ == "__main__":
 
             msg["payload"] = TotalApp.count
 
+            return msg
+        
+        @register("property-test", properties = [
+            InputProperty("input_prop", "input property"),
+            ListProperty("list_prop", [ "list", "property" ]),
+            DictProperty("dict_prop", { "dict": "property" }),
+            SpinnerProperty("spinner_prop", 1),
+            ComboBoxProperty("cbox_prop", [ "combobox", "property" ], "property")
+        ])
+        def property_test(props:dict, msg:dict) -> dict:
+            print(props, msg)
             return msg
         
         def test2(self, pros:dict, msg:dict) -> dict: