+++ tests/standalone_test.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
 import os
-from noderedpy import NodeProperty, RED, StandaloneServer
+from noderedpy import (
+    RED, StandaloneServer,
+    InputProperty, ListProperty, DictProperty
+)
 from noderedpy.decorator import register
 
 __dirname = os.path.dirname(os.path.realpath(__file__))
@@ -9,9 +12,9 @@ if __name__ == "__main__":
         count = 0
 
         @register("test", properties = [
-            NodeProperty("test_prop", "str", "1234"),
-            NodeProperty("list_prop", "list", [ 1, 2, 3, 4 ]),
-            NodeProperty("dict_prop", "dict", { "a": 1 })
+            InputProperty("test_prop", "1234"),
+            ListProperty("list_prop", [ 1, 2, 3, 4 ]),
+            DictProperty("dict_prop", { "a": 1 })
         ])
         def test(props:dict, msg:dict) -> dict:
             TotalApp.count += 1