+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 ## version 0.1.0
 - Initial Release
 
+<br/>
+
 ## version 0.1.1
 - type support for "list" and "dict"
 - add "display_icon" parameter to "NodeProperty"
   - if None, set icon by "type" of property
+
+<br/>
+
+## version 0.1.2
+- divide "NodeProperty" to multiple Property classes
+  - for input, InputProperty
+  - for list, ListProperty
+  - for dict(json), DictProperty
+  - for spinner, SpinnerProperty
+  - for combobox, ComboBoxProperty